feat: integrate ongoing training
