+++ assets/js/game.js
@@ -95,6 +95,102 @@ function neuralDecision(net, inputs) {
     return { angle, power, weapon };
 }
 
+// ==== Training Functions ====
+const POPULATION = 20;
+const MUTATION_RATE = 0.1;
+const MUTATION_STRENGTH = 0.5;
+
+function cloneNet(net) {
+    return JSON.parse(JSON.stringify(net));
+}
+
+function mutate(net) {
+    const mutateArray = (arr) => {
+        for (let i = 0; i < arr.length; i++) {
+            if (Array.isArray(arr[i])) mutateArray(arr[i]);
+            else if (Math.random() < MUTATION_RATE)
+                arr[i] += (Math.random() * 2 - 1) * MUTATION_STRENGTH;
+        }
+    };
+    mutateArray(net.hiddenWeights);
+    mutateArray(net.hiddenBias);
+    mutateArray(net.outputWeights);
+    mutateArray(net.outputBias);
+}
+
+function simulateShot(net, enemyX) {
+    const inputs = [enemyX / 800, 0, 0, 0, Math.abs(enemyX) / 800];
+    const { angle, power, weapon } = neuralDecision(net, inputs);
+    const rad = (angle * Math.PI) / 180;
+    const w = WEAPONS[weapon];
+    const speed = w.speed * power;
+    const vx = Math.cos(rad) * speed;
+    const vy = Math.sin(rad) * speed;
+    const t = (vy * 2) / GRAVITY;
+    const landingX = vx * t;
+    const dist = Math.abs(landingX - enemyX);
+    let damage = 0;
+    if (dist <= w.radius) {
+        damage = w.damage * (1 - dist / w.radius);
+    }
+    const closeness = Math.max(0, (w.radius - dist) / w.radius);
+    return damage + closeness * 10;
+}
+
+function evaluate(net) {
+    let fitness = 0;
+    for (let i = 0; i < 5; i++) {
+        const enemyX = 300 + Math.random() * 200;
+        fitness += simulateShot(net, enemyX);
+    }
+    return fitness / 5;
+}
+
+function evolveStep(population) {
+    const scored = population.map((net) => ({ net, score: evaluate(net) }));
+    scored.sort((a, b) => b.score - a.score);
+    const best = scored[0].net;
+    const survivors = scored.slice(0, POPULATION / 2).map((s) => s.net);
+    const nextPop = survivors.map(cloneNet);
+    while (nextPop.length < POPULATION) {
+        const parent = cloneNet(
+            survivors[Math.floor(Math.random() * survivors.length)],
+        );
+        mutate(parent);
+        nextPop.push(parent);
+    }
+    return { best, nextPop, score: scored[0].score };
+}
+
+function evolve(populationSize = POPULATION, generations = 50) {
+    let population = Array.from({ length: populationSize }, createRandomNet);
+    let best = population[0];
+    for (let g = 0; g < generations; g++) {
+        const result = evolveStep(population);
+        population = result.nextPop;
+        best = result.best;
+        console.log(
+            `Generation ${g + 1}: best fitness = ${result.score.toFixed(2)}`,
+        );
+    }
+    return best;
+}
+
+let trainingPopulation = null;
+let trainingGen = 0;
+
+function startBackgroundTraining() {
+    trainingPopulation = Array.from({ length: POPULATION }, createRandomNet);
+    if (trainedNet) trainingPopulation[0] = cloneNet(trainedNet);
+    setInterval(() => {
+        const result = evolveStep(trainingPopulation);
+        trainingPopulation = result.nextPop;
+        trainedNet = cloneNet(result.best);
+        trainingGen++;
+        console.log(`Background generation ${trainingGen} complete`);
+    }, 1000);
+}
+
 let canvas,
     ctx,
     terrain,
@@ -476,9 +572,14 @@ if (typeof module !== 'undefined') {
         neuralDecision,
         GRAVITY,
         loadTrainedNet,
+        evolve,
+        startBackgroundTraining,
     };
 } else {
     window.onload = () => {
-        loadTrainedNet().then(init);
+        loadTrainedNet().then(() => {
+            init();
+            startBackgroundTraining();
+        });
     };
 }
